Fixes 001

+++ src/main/java/kg/attractor/jobsearch_remake/dto/ResumeDto.java
@@ -8,7 +8,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -20,7 +19,7 @@ public class ResumeDto {
     private Integer id;
 
     @NotNull(message = "Applicant is required")
-    private Integer applicantId;
+    private Long applicantId;
 
     @NotBlank(message = "Resume name cannot be blank")
     private String name;
